bunch of fixes
sign-in, register only for anonymous users; Long -> Double points; removed long names in queries in repositories, now: imported names; not able to post predictions twice; better looking text inputs; case when registration input fields are empty; numerated classification; some margin top; delete DELETE/SELECT statements from CREATE TABLE query

+++ App/TyperF1/src/main/java/com/typerf1/typerf1/dto/grandprix/GrandPrixScore.java
@@ -18,9 +18,9 @@ public class GrandPrixScore {
     String sessionName;
     String participantName;
     String participantSurname;
-    Long pointsSum;
+    Double pointsSum;
 
-    public GrandPrixScore(String grandPrixName, String participantName, String participantSurname, Long pointsSum) {
+    public GrandPrixScore(String grandPrixName, String participantName, String participantSurname, Double pointsSum) {
         this.grandPrixName = grandPrixName;
         this.participantName = participantName;
         this.participantSurname = participantSurname;
